@@ -1,95 +1,51 @@
 {
-  "name": "black-dashboard-angular",
-  "version": "1.0.0",
-  "description": "Black Dasboard Angular, a Creative Tim Desgin. Coded by Creative Tim",
+  "name": "angular-webshop",
+  "version": "0.0.0",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/creativetimofficial/black-dashboard-angular.git"
-  },
-  "keywords": [
-    "angular dashboard",
-    "angular admin",
-    "angular bootstrap admin",
-    "angular bootstrap dashboard",
-    "angular panel",
-    "angular bootstrap template",
-    "angular admin template",
-    "angular free",
-    "angular free dashboard",
-    "angular free bootstrap dashboard",
-    "free admin template",
-    "angular admin",
-    "angular admin panel",
-    "angular dashboard",
-    "angular dashboard free",
-    "angular admin free",
-    "angular free admin panel",
-    "angular freebie",
-    "angular freebie dashboard",
-    "angular freebie bootstrap dashboard",
-    "freebie admin template",
-    "angular dashboard freebie",
-    "angular admin freebie",
-    "angular freebie admin panel",
-    "creative tim",
-    "web dashboard",
-    "bootstrap 4 dashboard",
-    "bootstrap 4",
-    "css3 dashboard",
-    "bootstrap 4 admin",
-    "black dashboard bootstrap 4",
-    "frontend",
-    "responsive bootstrap 4 dashboard",
-    "free dashboard",
-    " free admin dashboard",
-    "free bootstrap 4 admin dashboard"
-  ],
-  "author": "Creative Tim <hello@creative-tim.com> (https://www.creative-tim.com/)",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/creativetimofficial/black-dashboard-angular/issues"
-  },
-  "homepage": "https://demos.creative-tim.com/black-dashboard-angular/#/dashboard",
-  "private": false,
+  "private": true,
   "dependencies": {
-    "@angular/animations": "8.0.2",
-    "@angular/common": "8.0.2",
-    "@angular/compiler": "8.0.2",
-    "@angular/core": "8.0.2",
-    "@angular/forms": "8.0.2",
-    "@angular/platform-browser": "8.0.2",
-    "@angular/platform-browser-dynamic": "8.0.2",
-    "@angular/router": "8.0.2",
-    "@ng-bootstrap/ng-bootstrap": "4.2.1",
-    "chart.js": "2.8.0",
-    "core-js": "3.1.4",
-    "ngx-toastr": "10.0.4",
-    "rxjs": "6.5.2",
-    "tslib": "1.10.0",
-    "zone.js": "0.9.1"
+    "@angular/animations": "~8.2.4",
+    "@angular/common": "~8.2.4",
+    "@angular/compiler": "~8.2.4",
+    "@angular/core": "~8.2.4",
+    "@angular/forms": "~8.2.4",
+    "@angular/platform-browser": "~8.2.4",
+    "@angular/platform-browser-dynamic": "~8.2.4",
+    "@angular/router": "~8.2.4",
+    "angular-font-awesome": "^3.1.2",
+    "bootstrap": "^4.3.1",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.800.3",
-    "@angular/cli": "8.0.3",
-    "@angular/compiler-cli": "8.0.2",
-    "@angular/language-service": "8.0.2",
-    "@types/node": "12.0.8",
-    "@types/jasmine": "3.3.13",
-    "@types/jasminewd2": "2.0.6",
-    "codelyzer": "5.1.0",
-    "jasmine-core": "3.4.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.1.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.5",
-    "karma-jasmine": "2.0.1",
-    "karma-jasmine-html-reporter": "1.4.2",
-    "protractor": "5.4.2",
-    "ts-node": "8.3.0",
-    "tslint": "5.17.0",
-    "typescript": "3.4.5"
+    "@angular-devkit/build-angular": "~0.803.2",
+    "@angular/cli": "~8.3.2",
+    "@angular/compiler-cli": "~8.2.4",
+    "@angular/language-service": "~8.2.4",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.0",
+    "eslint": "^6.4.0",
+    "eslint-plugin-angular": "^4.0.1",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
