@@ -1,31 +1,45 @@
-import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { NgModule } from "@angular/core";
-import { FormsModule } from "@angular/forms";
-import { HttpClientModule } from "@angular/common/http";
-import { RouterModule } from "@angular/router";
-import { ToastrModule } from 'ngx-toastr';
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
+import { FormsModule } from '@angular/forms';
 
-import { AppComponent } from "./app.component";
-import { AdminLayoutComponent } from "./layouts/admin-layout/admin-layout.component";
-import { AuthLayoutComponent } from './layouts/auth-layout/auth-layout.component';
-
-import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
-
-import { AppRoutingModule } from "./app-routing.module";
-import { ComponentsModule } from "./components/components.module";
+import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+import { CreateOrdersComponent } from './page/create-orders/create-orders.component';
+import { CreateUsersComponent } from './page/create-users/create-users.component';
+import { NavComponent } from './nav/nav.component';
+import { OrdersComponent } from './page/orders/orders.component';
+import { UpdateOrdersComponent } from './page/update-orders/update-orders.component';
+import { UpdateUsersComponent } from './page/update-users/update-users.component';
+import { UsersComponent } from './page/users/users.component';
+import { ProjectAddComponent } from './project-add/project-add.component';
+import { ProjectEditComponent } from './project-edit/project-edit.component'
+import { ProjectsComponent } from './projects/projects.component';
+import { IndexComponent } from './page/index/index.component';
+import { SearchPipe } from './pipe/search.pipe';
 
 @NgModule({
-  imports: [
-    BrowserAnimationsModule,
-    FormsModule,
-    HttpClientModule,
-    ComponentsModule,
-    NgbModule,
-    RouterModule,
-    AppRoutingModule,
-    ToastrModule.forRoot()
+  declarations: [
+    AppComponent,
+    OrdersComponent,
+    ProjectsComponent,
+    ProjectAddComponent,
+    ProjectEditComponent,
+    UpdateOrdersComponent,
+    CreateOrdersComponent,
+    UsersComponent,
+    UpdateUsersComponent,
+    CreateUsersComponent,
+    NavComponent,
+    IndexComponent,
+    SearchPipe,
   ],
-  declarations: [AppComponent, AdminLayoutComponent, AuthLayoutComponent],
+  imports: [
+    BrowserModule,
+    AppRoutingModule,
+    HttpClientModule,
+    FormsModule
+  ],
   providers: [],
   bootstrap: [AppComponent]
 })
