@@ -1,28 +1,23 @@
-import { AppPage } from "./app.po";
-import { browser, logging } from "protractor";
+import { AppPage } from './app.po';
+import { browser, logging } from 'protractor';
 
-describe("workspace-project App", () => {
+describe('workspace-project App', () => {
   let page: AppPage;
 
   beforeEach(() => {
     page = new AppPage();
   });
 
-  it("should display welcome message", () => {
+  it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getTitleText()).toEqual("Welcome to black-dashboard-angular!");
+    expect(page.getTitleText()).toEqual('angular-webshop app is running!');
   });
 
   afterEach(async () => {
     // Assert that there are no errors emitted from the browser
-    const logs = await browser
-      .manage()
-      .logs()
-      .get(logging.Type.BROWSER);
-    expect(logs).not.toContain(
-      jasmine.objectContaining({
-        level: logging.Level.SEVERE
-      } as logging.Entry)
-    );
+    const logs = await browser.manage().logs().get(logging.Type.BROWSER);
+    expect(logs).not.toContain(jasmine.objectContaining({
+      level: logging.Level.SEVERE,
+    } as logging.Entry));
   });
 });
