@@ -7,7 +7,7 @@ module.exports = class WhereGenerator {
   /**
    * Generates a MySQL WHERE syntax based on the url query string.
    * @param {req.query} queryObject request query object from the request url.
-   * @returns {string} A MySQL compatible "WHERE cond1=val1 AND cond2=val2" string. 
+   * @returns {string} A MySQL compatible "WHERE cond1=val1 AND cond2=val2" string.
    */
   getWhereString(queryObject) {
     this._emptyPreviousQuery();
@@ -24,10 +24,10 @@ module.exports = class WhereGenerator {
     this.queryKeys.forEach((key, index) => {
       if (index !== 0) {
         this.whereString = this.whereString
-          .concat(` AND `);
+          .concat(' AND ');
       }
       this.whereString = this.whereString
-        .concat(`${key} = ${this.query[key]}`);
+        .concat(`${key} = ${this._apostropheByType(this.query[key])}`);
     });
   }
 
@@ -43,4 +43,13 @@ module.exports = class WhereGenerator {
     this.queryKeys = Object.keys(this.query);
   }
 
-}
+  _apostropheByType(value) {
+    if (typeof value === 'string') {
+      return `'${value}'`;
+    }
+    if (typeof value === 'number') {
+      return `${value}`;
+    }
+  }
+
+};
